Правки после код ревью
