--- src/main/java/ru/practicum/shareit/item/ItemService.java
@@ -1,5 +1,0 @@
-package ru.practicum.shareit.item;
-
-public interface ItemService extends BaseItemService {
-    void deleteItem(Long id, Long ownerId);
-}
