--- src/main/java/ru/practicum/shareit/user/UserService.java
@@ -1,5 +1,0 @@
-package ru.practicum.shareit.user;
-
-public interface UserService extends BaseUserService {
-    void deleteUser(Long id);
-}
